moving getUserRegistrations to registration file
